+++ RELEASE-NOTES.md
@@ -22,6 +22,7 @@
 * Add support for passing OAuth token via `--token` flag.
 * Added feature to allow suppressing the new version warning message in the Snowflake CLI, configurable via a configuration file or environment variable.
 * Add support for `--format=CSV`
+* Fix JSON serialization for Decimal
 
 ## Fixes and improvements
 * Fixed failing snow sql command when snowflake.yml is invalid and query has no templating.